Separated Release from Metrcis before check
